Moved Role and the associated classes into clouddriver-aws.
