+++ clouddriver-aws/src/main/groovy/com/netflix/spinnaker/clouddriver/aws/model/Role.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.netflix.spinnaker.clouddriver.model;
+package com.netflix.spinnaker.clouddriver.aws.model;
 
 import java.util.Collection;
