+++ clouddriver-ecs/src/main/java/com/netflix/spinnaker/clouddriver/ecs/cache/model/IamRole.java
@@ -16,8 +16,8 @@
 
 package com.netflix.spinnaker.clouddriver.ecs.cache.model;
 
-import com.netflix.spinnaker.clouddriver.model.Role;
-import com.netflix.spinnaker.clouddriver.model.TrustRelationship;
+import com.netflix.spinnaker.clouddriver.aws.model.Role;
+import com.netflix.spinnaker.clouddriver.aws.model.TrustRelationship;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
